+++ CHANGELOG.md
@@ -1,3 +1,10 @@
 ## 1.0.0 (January 15, 2016)
 
 - Initial release
+
+
+## 1.0.1 (January 27, 2016)
+
+- Add custom SimplePay checkout image
+- Fix paths
+- Hide "simplepay_transaction_id" from WooCommerce error message in checkout page